move the config to agent
