+++ internal/otelcollector/config/metric/agent/drop_istio_internal_metrics_test.go
@@ -1,4 +1,4 @@
-package gateway
+package agent
 
 import (
 	"testing"
