+++ internal/otelcollector/config/metric/agent/processors.go
@@ -26,6 +26,7 @@ func makeProcessorsConfig(inputs inputSources) Processors {
 
 		if inputs.istio {
 			processorsConfig.InsertInputSourceIstio = makeEmittedByConfig(metric.InputSourceIstio)
+			processorsConfig.DropIstioMetricsToTelemetryComponents = makeFilterToDropMetricsForTelemetryComponents()
 		}
 	}
