+++ internal/otelcollector/config/metric/gateway/config_builder.go
@@ -91,17 +91,17 @@ func makeServiceConfig() config.Service {
 // declareComponentsForMetricPipeline enriches a Config (exporters, processors, etc.) with components for a given telemetryv1alpha1.MetricPipeline.
 func declareComponentsForMetricPipeline(ctx context.Context, otlpExporterBuilder *otlpexporter.ConfigBuilder, pipeline *telemetryv1alpha1.MetricPipeline, cfg *Config, envVars otlpexporter.EnvVars) error {
 	declareDropInputFilters(pipeline, cfg)
-	declareDropIstioInternalMetrics(pipeline, cfg)
+	//declareDropIstioInternalMetrics(pipeline, cfg)
 	declareNamespaceFilters(pipeline, cfg)
 	return declareOTLPExporter(ctx, otlpExporterBuilder, pipeline, cfg, envVars)
 }
 
-func declareDropIstioInternalMetrics(pipeline *telemetryv1alpha1.MetricPipeline, cfg *Config) {
-	input := pipeline.Spec.Input
-	if isIstioInputEnabled(input) {
-		cfg.Processors.DropIstioMetricsToTelemetryComponents = makeFilterToDropMetricsForTelemetryComponents()
-	}
-}
+//func declareDropIstioInternalMetrics(pipeline *telemetryv1alpha1.MetricPipeline, cfg *Config) {
+//	input := pipeline.Spec.Input
+//	if isIstioInputEnabled(input) {
+//		cfg.Processors.DropIstioMetricsToTelemetryComponents = agent.makeFilterToDropMetricsForTelemetryComponents()
+//	}
+//}
 
 func declareDropInputFilters(pipeline *telemetryv1alpha1.MetricPipeline, cfg *Config) {
 	input := pipeline.Spec.Input
@@ -166,7 +166,7 @@ func makeServicePipelineConfig(pipeline *telemetryv1alpha1.MetricPipeline) confi
 
 	processors = append(processors, makeDropInputFilters(input)...)
 
-	processors = append(processors, makeDropIstioInternalMetrics(input)...)
+	//processors = append(processors, makeDropIstioInternalMetrics(input)...)
 
 	if isRuntimeInputEnabled(input) && shouldFilterByNamespace(input.Runtime.Namespaces) {
 		processors = append(processors, makeNamespaceFilterID(pipeline.Name, metric.InputSourceRuntime))
@@ -190,13 +190,13 @@ func makeServicePipelineConfig(pipeline *telemetryv1alpha1.MetricPipeline) confi
 	}
 }
 
-func makeDropIstioInternalMetrics(input telemetryv1alpha1.MetricPipelineInput) []string {
-	var processors []string
-	if isIstioInputEnabled(input) {
-		processors = append(processors, "filter/drop-istio-metrics-to-internal-components")
-	}
-	return processors
-}
+//func makeDropIstioInternalMetrics(input telemetryv1alpha1.MetricPipelineInput) []string {
+//	var processors []string
+//	if isIstioInputEnabled(input) {
+//		processors = append(processors, "filter/drop-istio-metrics-to-internal-components")
+//	}
+//	return processors
+//}
 
 func makeDropInputFilters(input telemetryv1alpha1.MetricPipelineInput) []string {
 	var processors []string
